requirements.txt
discord.py==1.6.0
google-cloud-secret-manager==2.2.0